create test data in a new file
